requirements.txt
appdirs==1.4.3
beautifulsoup4==4.8.2
certifi==2019.11.28
cfgv==3.1.0
chardet==3.0.4
click==7.1.1
distlib==0.3.0
filelock==3.0.12
futures==3.1.1
identify==1.4.13
idna==2.9
importlib-metadata==1.6.0
importlib-resources==1.4.0
lxml==4.5.0
nodeenv==1.3.5
oauthlib==3.1.0
packaging==20.3
Pillow==7.1.1
pluggy==0.13.1
pre-commit==2.2.0
py==1.8.1
pyglet==1.5.2
PyJWT==1.7.1
pyparsing==2.4.6
PyPDF2==1.26.0
PyQt5==5.11.3
PyQt5-sip==4.19.13
PySocks==1.7.1
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
soupsieve==2.0
toml==0.10.0
tox==3.14.6
translate==3.5.0
translator==0.0.9
tweepy==3.8.0
twilio==6.38.0
urllib3==1.25.8
virtualenv==20.0.15
zipp==3.1.0
